third_party/WebKit/LayoutTests/hittesting/hittest-child-of-inlineblock.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#roundedDiv {
    width: 200px;
    height: 200px;
    border-radius: 5px;
    border: 5px solid transparent;
    display: inline-block;
}
#roundedDivChild {
    width: 200px;
    height: 100px;
    margin-left: 400px;
}
</style>
<div id="roundedDiv">
<div id="roundedDivChild">Mouse over me!<br>The box should turn green.</div>
</div>
<div id="log"></div>
<script>
test(function(t)
{
    var element = document.getElementById("roundedDivChild");
    var x = element.offsetLeft + element.offsetWidth / 2;
    var y = element.offsetTop + element.offsetHeight / 2;

    var element = document.elementFromPoint(x, y);
    assert_equals(element.nodeName, 'DIV');
    assert_equals(element.id, 'roundedDivChild');
}, "elementFromPoint should return an element under a point");
</script>
